+++ app/src/main/java/woowacourse/payments/ui/component/registration/CardExpirationDateTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.component.registration
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
