+++ app/src/main/java/woowacourse/payments/ui/screen/cards/CardsScreen.kt
@@ -4,15 +4,11 @@ import android.app.Activity
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -22,7 +18,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
@@ -32,15 +27,16 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.R
 import woowacourse.payments.ui.common.getParcelableExtraCompat
-import woowacourse.payments.ui.component.CardsTopAppBar
 import woowacourse.payments.ui.component.PaymentCard
+import woowacourse.payments.ui.component.cards.CardsTopAppBar
+import woowacourse.payments.ui.component.cards.RegistrationBox
 import woowacourse.payments.ui.model.CardExpirationDateUiModel
 import woowacourse.payments.ui.model.CardNumberUiModel
 import woowacourse.payments.ui.model.CardUiModel
 import woowacourse.payments.ui.model.CardholderNameUiModel
 import woowacourse.payments.ui.screen.cards.CardsActivity.Companion.EXTRA_CARDS_REGISTER_NEW_CARD
 import woowacourse.payments.ui.screen.registration.CardRegistrationActivity
-import woowacourse.payments.ui.theme.TextGray
+import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 @Composable
 fun CardsScreen(
@@ -131,25 +127,6 @@ private fun RegistrationGuideText() {
     )
 }
 
-@Composable
-private fun RegistrationBox(onClickRegistration: () -> Unit) {
-    Box(
-        modifier =
-            Modifier
-                .size(width = 208.dp, height = 124.dp)
-                .background(Color.LightGray)
-                .clickable { onClickRegistration() },
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            text = stringResource(R.string.cards_screen_registration_symbol),
-            fontSize = 34.sp,
-            fontWeight = FontWeight.Bold,
-            color = TextGray,
-        )
-    }
-}
-
 @Preview(showBackground = true, name = "데이터 존재 X")
 @Composable
 private fun NoContentPreview() {
@@ -173,9 +150,11 @@ private fun HasOneContentPreview() {
             ),
         )
 
-    CardsScreen(
-        cardsScreenUiState = uiState,
-    )
+    AndroidpaymentsTheme {
+        CardsScreen(
+            cardsScreenUiState = uiState,
+        )
+    }
 }
 
 @Preview(showBackground = true, name = "데이터 2개 이상 존재")
@@ -196,7 +175,9 @@ private fun HasMultipleContentPreview() {
                 ),
             ),
         )
-    CardsScreen(
-        cardsScreenUiState = uiState,
-    )
+    AndroidpaymentsTheme {
+        CardsScreen(
+            cardsScreenUiState = uiState,
+        )
+    }
 }
